requirements.txt
aniso8601==9.0.1
asttokens==2.4.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
fonttools==4.54.1
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mysql-connector==2.2.9
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
php==1.2.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sty==1.0.6
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
